requirements.txt
pydantic==2.10.6
requests==2.32.3
pip==25.0
attrs==23.1.0
wheel==0.45.1
filelock==3.17.0
pytest==8.3.4
typing_extensions==4.12.2
pluggy==1.5.0
exceptiongroup==1.2.2
iniconfig==2.0.0
argcomplete==3.1.1
allure-python-commons==2.13.5
packaging==23.2
setuptools==65.5.0
Jinja2==3.1.2
allure-pytest==2.13.5
Faker==35.0.0
structlog==25.1.0
curlify==2.2.1